Fixed sample pages
Forgot  I moved the location of the sidebar in the code so that the SCSS could nest properly.

The aside tag was removed and is now just an nav tag. All my research says if this is a menu then it should be a nav element.

So the modified html places the sidebar before the header navbar as the SCSS transitions need to control the header, content area, and footer.