Migrated to Bootstrap 5.3 for themes
I have manually added BS 5.3 to the project. This is a first run at it and there may need to be more adjustments.

I also added a toggle switch in the header menu so you can toggle between the light and dark themes. It will also store it to the browsers local storage so that it will remember your last selection. Should the setting no longer be in storage then it will look to your browsers to try and get it from the browser preference.

The dist folder reflects all these changes.